Move GcpSecretLookup to plugin_utils because may not be ansible module on the remote host
